Refactor code by separating BookManager and PersonManager into their own files
