+++ book_manager.rb
@@ -1,0 +1,31 @@
+# Class responsible for book management
+class BookManager
+    attr_accessor :books
+  
+    def initialize
+      @books = []
+    end
+  
+    def list_all_books
+      if @books.empty?
+        puts 'No books available.'
+      else
+        puts 'List of all books:'
+        @books.each_with_index do |book, index|
+          puts "#{index}) Title: \"#{book.title}\", Author: #{book.author}"
+        end
+      end
+    end
+  
+    def create_book
+      print 'Title: '
+      title = gets.chomp
+      print 'Author: '
+      author = gets.chomp
+  
+      new_book = Book.new(title, author)
+      @books << new_book
+      puts 'Book created successfully'
+    end
+  end
+
